fixup: move classes to XdsResolver
